Added utils folder and contrast file with functions for determining color contrast
